start on compositing this
