+++ redbot/cogs/mod/__init__.py
@@ -1,6 +1,13 @@
 from redbot.core.bot import Red
-from .mod import Mod
+from .modcore import Mod
+
+from .bans import KickBanMixin
+from .mutes import MuteMixin
+from .misc import MiscMixin
+from .utils import UtilsMixin
 
 
-def setup(bot: Red):
+async def setup(bot: Red):
+    cog = Mod(bot)
+    await cog._casetype_registration()
     bot.add_cog(Mod(bot))
