+++ redbot/cogs/mod/log.py
@@ -1,3 +1,4 @@
+# Why was this already in it's own file? Move to utils entirely?
 import logging
